Updated release script

+++ release/build_and_push.sh
@@ -1,5 +1,15 @@
 #!/bin/bash
-TAG_NAME=therealbfg/git_httpd_php
+
+SCRIPT_DIR="$(dirname "$(readlink -f "$0")")"
+
+if [ ! -f $SCRIPT_DIR/TAG_NAME ]
+then
+  echo "You need to write the name of the tag in '$SCRIPT_DIR/TAG_NAME'"
+  exit 1
+fi
+
+TAG_NAME=$(cat $SCRIPT_DIR/TAG_NAME)
+
 fname=$(mktemp)
 
 docker build --pull . | tee $fname
